FIX: newly created tiles had incorrect price. REMOVE: getTileAt(). ENHANCEMENT: describeTile() returns id of "0x0" if the tile does not exist. ENHANCEMENT: Remove redundant Tile reference lookups among private functions by passing Tile directly. ENHANCEMENT: Remove redundant height checking, unpacking logic for private functions that operate on known valid blocks.

+++ sol/UpcityBase.sol
@@ -114,7 +114,7 @@ contract UpcityBase {
 	// The ratio of collected resources to share with neighbors, in ppm.
 	uint64 internal constant TAX_RATE = $$(TO_PPM(1/NUM_NEIGHBORS));
 	// The minimum tile price.
-	uint256 internal constant MINIMUM_TILE_PRICE = $$(int(ONE_TOKEN * MINIMUM_TILE_PRICE));
+	uint256 internal constant MINIMUM_TILE_PRICE = $$(uint256(ONE_TOKEN * MINIMUM_TILE_PRICE));
 	// How much to increase the base tile price every time it's bought, in ppm.
 	uint64 internal constant PURCHASE_MARKUP = $$(TO_PPM(1+PURCHASE_MARKUP));
 	// The number of seasons.
